Massive Spawner merge method rework
Changed a lot on how the spawners will absorb, and also the timing system is more or less gone. also removed all uneeede/commented code. This should make things neater. system now absorbs spawns inside the spawn method intead of outside. this means that when a spawn is called for a spawner, it will absorb in the method before actualyl spawning.