Sizes Adjusted
Sizes for some of the sprite imports have been adjusted a bit. a minimap as been introduced. The minimap should be ignored as it is in the back layer aslong witht he camera that is previewing it. the minimap will probably be moved somwhere else or removed in it's entirety at a later date.