Multiple Changes
-Added a loading screen to cover while level generates.
-More Size Changes to the sprite imports. Orthographic size chagned from 60 to 24
-Changed where instantiated GameObjects go, they will no longer spawned unparented, they will now be parented where they belong when created for better organization.
-PlayerPoint on the map generator can now spawn rooms. This will allow us to use the PlayerPoint on the map to keep track of where the player is going and spawn rooms accordingly.

NOTE: as of now there is no real system to keep track of what rooms have spawned/visited and which have not. I will be coming up with a system to spawn rooms with the PlayerPoint and then assingn those spawned rooms to the corresponding DataPoint that caused the initial spawn, that way when the PlayerPoint re-enters an already visited/spawned DataPoint it will not spawn another room. Still not sure what rooms will do when they are offscreen, debating on leaving them be, or disabling them while they are off screen for resource saving and having them re-enable upon the player entering again.