Open and Char_Select
I have finished the Opening_Scene. Only thing that needs to be added is animation and a Title.

As for the character Selection Scene, it needs alot more animations and some sort of functioning UI.

Currently Im focusing on how the character would animate when moving.
I will create a loading screen later.