Room transition/movement
Added the ability to move between rooms.
Rooms will now Spawn doors connecting the rooms.
Rooms once spawned (they only spawn when entered for the first time) will not spawn another room upon revisting it.
Found that one of the template rooms was a bit off and caused an issue when moving between rooms, this was fixed and not an issue now.

NOTE: after left enough rooms from spawn, you will noticec gizmos overlapping the camera/scene, this is not an error or bug, this is due to how Unity renders gizmos. They will disappear once disabled or set to 3D. This was left in there since it isn't an actual problem and since the minimap will proabably be removed at somepoint, it is only used now for debuggin purposes.

TO DO NOTE: I must start to consider making the generation system modular/independent, as of now a lot of the system relies on the Start function. I would like to use a GameHandler/GameManager game object that has a script that can trigger/cause these different systems, like map generation, room generation, etc. This will make it easier moving on in the future, hopefully i can get this done at some point.