Changed JoyStick Controls
-Made major changes to the joystick controls. instead of now reading Mouse inputs it will read Touch Inputs. This system required a new script that managed all touch inputs and passed relavent information over the PlayerJoystick script for use.
-A Attack joystick is in need of finishing but the base is there to just copy and move the PlayerJoystick script over with a few changes to only work for the gun.
-TouchInput Script still needs to be adjusted to check for specific collisions, right now it detects and looks for a touch where there is no collision at all.

NOTE: Since the Joystick has been reworked to incorporate touch controls and no longer works with mouse, you must use a touch device when testing to be able to move. This means either use Unity and test on a Touchscreen enabled laptop OR export the build and run the APK on any touchscreen device for testing. YES, this must be done EVERYTIME, it is very annoying.
ALSO OF NOTE: finally moved the build over to android, no longer is it on PC. This means our game can run on android now. something smaller to mention, made some of the assets scale to the screen, specifically the UI and character select menu. changed JoyStick Sizes around as I looked for optimal size, I think i found it. Still need to mess with optimal speed.