Copy Branch
This branch contains changes and implementations I made before put into the WIP branch to no interefere too much with the main code. What was added was powerups and music to the dungeon. Still Working on it. But pushing in the code so I can go back to the other branches and add UI, and pause button.