Added SlimeKing
Got the SlimeKing to work along with his splitting mechanic upon death.
-Added a new Room that houses the new Slime King. Still no proper "BossRoom" setup as his new room spawns in as a regular room.
-Added player death & health counter.
The player can now die and upon death will bring the player to the death screen where they can tap to restart and select a new character.
NOTE: The player character's health can be changed in the inspector. Right now it is 10 so everything can 1 shot them, this was for testing.
I also changed hitbox of the KingSlime, it is made of a unique polygon that loosly traces the shape of the KingSlime, the only issue is that the shape of the slime changes when it's sprite changes, this could be fixed through the animation, but it is small and I'm lazy.
With this change I added the death screen to the scene loader script and to the list of loaded scene upon launching the game.
PS: Unable to Add boss Healthbar as Slider component might be broken/bugged with my Unity version, I had a hard time trying to access the Slider component in C#, might look at it later.